Partially-implemented signature fragments.
- Renamed `SigningKey` to `PrivateKey`.
- Implemented `PrivateKey.create_signature`.
- Automatically attach balance and key_index to addresses in
``getInputs`` and ``getBalances``.
- Moved transaction-related classes into own package.
- Refactored magic number 3.
- Renamed ``as_json`` to ``as_json_compatible``.
