+++ iota/json.py
@@ -10,6 +10,6 @@ from json.encoder import JSONEncoder as BaseJsonEncoder
 class JsonEncoder(BaseJsonEncoder):
   """JSON encoder with support for custom types."""
   def default(self, o):
-    if hasattr(o, 'as_json'):
-      return o.as_json()
+    if hasattr(o, 'as_json_compatible'):
+      return o.as_json_compatible()
     return super(JsonEncoder, self).default(o)
