+++ test/crypto/signing_test.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import, division, print_function, \
 from unittest import TestCase
 
 from iota.crypto.signing import KeyGenerator
-from iota.crypto.types import SigningKey
+from iota.crypto.types import PrivateKey
 
 
 # noinspection SpellCheckingInspection
@@ -40,7 +40,7 @@ class KeyGeneratorTestCase(TestCase):
       # Note that the result is always a list, even when generating a
       # single key.
       [
-        SigningKey(
+        PrivateKey(
           b'BWFTZWBZVFOSQYHQFXOPYTZ9SWB9RYYHBOUA9NOYSWGALF9MSVNEDW9A9FLGBRWKED'
           b'MPEIPRKBMRXRLLFJCAGVIMXPISRGXIJQ9BOBHKJEUKDEUUWYXJGCGAWHYBQHBPMRTZ'
           b'FPBGNLMKPZYXZPXFSPFUWZNRWYXUEWMP9URKVVJOSWEPJKSMPLWZPIZGOTVAA9QQOC'
@@ -83,7 +83,7 @@ class KeyGeneratorTestCase(TestCase):
       kg.get_keys(start=1),
 
       [
-        SigningKey(
+        PrivateKey(
           b'WEVFRAOIRJBSBKQWF9JOTQXWUDLOIJRAC9WOOJNW99UAXVMUMSCMAABBXI99PRTAQL'
           b'UWJKVMM9DPSZSU9SAUN9URDWGGXIHWJJCDBAY9OQMQURNHZBD9E9CGERZC9RSUQVMZ'
           b'VYUTYXLH9CCEQPQLVDICQD9UCH9RPP9NSNZEBYERXHDEBEZUZOKNNEMXBYZPAYIIOB'
@@ -130,7 +130,7 @@ class KeyGeneratorTestCase(TestCase):
       kg.get_keys(start=13),
 
       [
-        SigningKey(
+        PrivateKey(
           b'ZIGSOMJRQUXQHMNP9NCEGWCNXXVMJW9BXYRTMVUWRVTFQ9GCMJOOENTBSJDKPQTWML'
           b'FGEMPNODQWJ9BIZJSFWOOQNYLHIGUAJJXGISEMZKVPQOLQIMKESDECLJLDJFTRSBCQ'
           b'UEAJTCMUDYYWWENQZAPI9B9B9RBOCEAUAQHQNCZTWDOCYCRXZNHYTTTTNUNROHJGCF'
@@ -171,7 +171,7 @@ class KeyGeneratorTestCase(TestCase):
 
   def test_get_keys_long_seed(self):
     """
-    Generating a SigningKey from a seed longer than 1 hash.
+    Generating a PrivateKey from a seed longer than 1 hash.
 
     This catches a regression caused by
     :py:meth:`iota.crypto.pycurl.Curl.squeeze` processing the wrong
@@ -189,7 +189,7 @@ class KeyGeneratorTestCase(TestCase):
       kg.get_keys(start=0),
 
       [
-        SigningKey(
+        PrivateKey(
           b'V9ZUOXOFKFSYMYRUKLXFTMLKBTDGXJWZZSEQINISGCXCPDQ9MFLTIEXXOEKPJPU9TW'
           b'WQAOWDNBVHODJSMTULXCDKTBPEBQMWNXRUWY9WPOFQZDOLYXWEABWU9DNFIUEETSHR'
           b'RNXXAGENAIQQYDBWIVXFZIKLNSEHVNKZGKJPAFTICTRNXCRDBXBEKBQF9XKS9LSVZU'
@@ -240,7 +240,7 @@ class KeyGeneratorTestCase(TestCase):
       kg.get_keys(start=1, count=2),
 
       [
-        SigningKey(
+        PrivateKey(
           b'VBGOUVYVPMMVMNYIASWCSDOECOVM9BLBAWECSHKURDUVTFBRMUCIADFORQXPPCDVGM'
           b'QAGEPRS9IVKLZTJOVOBFXWC9PDXBL9WCRYKCNNFWOAJCKKV9KZICCQKUOFUAUYTRJH'
           b'DTCRDLWCXKSPK9ZMVVFMGXBPQRCVAFMRAGVHWFKQEDPXPR9UTYOSKPMGMZZWLW9SBZ'
@@ -277,7 +277,7 @@ class KeyGeneratorTestCase(TestCase):
           b'REONEIFHT'
         ),
 
-        SigningKey(
+        PrivateKey(
           b'H9VZYRIEGPFTUOMXYHPDLGNNWEXKHOMNHI9YIEOCBXRQJTZMYW9LHXTZDQEYMWUAGY'
           b'SJASVKHVXRJBMZINYZX9WBUERAH9EV9XDESZESVMYMQRNCNERUUGCYLBKHZDDLAPGU'
           b'MIEMROYBKAIETO9JVWLFPOHJZTFOTU9FNMOECNZEOFG9DKVSQNQBXWLVDTBCPMRAUN'
@@ -365,7 +365,7 @@ class KeyGeneratorTestCase(TestCase):
       # This is the same as ``kg.get_keys(start=0, count=2)``, except
       # the order is reversed.
       [
-        SigningKey(
+        PrivateKey(
           b'ODJGKFVKDKETMOUH9OCDBIDQCEMEDVKMEOIKSDTIQFECONPDCZUITROKXYCTNEMVQI'
           b'KZWKIJQZVYEAEUFGWAZUJEQMZLBPPYLJYEBOEZNYDFWKLYB9S9GTVZOXGGQ9CVYVVX'
           b'ZNIPHGIARTYMUXOJVTQZSASWVK9CSRUJODBJRPCCDDFEPZRNYHIGMFRFXAFGPMACLZ'
@@ -402,7 +402,7 @@ class KeyGeneratorTestCase(TestCase):
           b'KVEPFJDTR'
         ),
 
-        SigningKey(
+        PrivateKey(
           b'TIOBFKKFAELHQOCLGJBGJZWVNSZMBPK9D9GTBZWDJBFFIKJWJBKIAAAAPDMYOCPDRN'
           b'BMK9QGJZAMIQZKSBNVVDLIVNRYVWQSHSJZZNGJGWNHGDMYCDZLIIUAPIPBKYSWIFNP'
           b'ZLVUGCN9MIQTYFYJWVIKXOUOFHAMCYZEVBHYABYPEQHXPHIPM9NWENCXRLUYQFILME'
@@ -456,7 +456,7 @@ class KeyGeneratorTestCase(TestCase):
 
       [
         # 2 iterations = key is twice as long!
-        SigningKey(
+        PrivateKey(
           b'VBGOUVYVPMMVMNYIASWCSDOECOVM9BLBAWECSHKURDUVTFBRMUCIADFORQXPPCDVGM'
           b'QAGEPRS9IVKLZTJOVOBFXWC9PDXBL9WCRYKCNNFWOAJCKKV9KZICCQKUOFUAUYTRJH'
           b'DTCRDLWCXKSPK9ZMVVFMGXBPQRCVAFMRAGVHWFKQEDPXPR9UTYOSKPMGMZZWLW9SBZ'
@@ -550,7 +550,7 @@ class KeyGeneratorTestCase(TestCase):
     self.assertEqual(
       next(generator),
 
-      SigningKey(
+      PrivateKey(
         b'LZDQHHPRICEESIHX9VVYCHLYSJDMXPLFOMMVOMUZKHSXLQYSQNDFHWLNKTCAJYPWUM'
         b'RMILEVQCHTGILDRABOOXPNRQHWBXVBRIZELVFEHTMYITSVBBBUQCZMDEUFGCTKHISF'
         b'MI9XQOBAIOVMCKNIGJQANHDMCWHYFJLDQLPMHLULLZZXYNZUNDLMRITVORDUEGNWKH'
@@ -591,7 +591,7 @@ class KeyGeneratorTestCase(TestCase):
     self.assertEqual(
       next(generator),
 
-      SigningKey(
+      PrivateKey(
         b'BTQKIPUTHFDMCKGVLBCFGEFLJHCHOANY9GBIRSODCOWTJGUWGFMDXWEB9HDNP9M9ZU'
         b'HVZQULQHEMXV9GMXDNYDHEJQZ9BSPIPHIVWNZDJWWDWONDKNVRFPZWNXZAVVTZCCMF'
         b'DBOYJZCGKIFJDKLMTCGGFUJAZ9FPOIBDSHDZCUYNVXCCDAFZLF9IMKZWTZFUMZSVCT'
@@ -642,7 +642,7 @@ class KeyGeneratorTestCase(TestCase):
     self.assertEqual(
       next(generator),
 
-      SigningKey(
+      PrivateKey(
         b'CBKFYOTDYHFWSIOUYAUAHQWOOQDNNQBTSSJPHREUWFBXZFYFHPHZJN9ILAMZYOXLBQ'
         b'MCN9T9F9VSCUSFVQDZROPWBKZ9ENENGFVDLMBNDOAAKZXVBXVAYJLSGVD9ZZSMDCRT'
         b'JUUJPTNRVYYTOALRFOIGAXHUIXWCMXEUGTH9UJKYWTWBEOVWPMNVHJSXXKBAUPYCLS'
@@ -683,7 +683,7 @@ class KeyGeneratorTestCase(TestCase):
     self.assertEqual(
       next(generator),
 
-      SigningKey(
+      PrivateKey(
         b'OTOFWEAUQFRSWBTJPPAACBIPCTMYAESBAGVMPVMH9IQAEEKEXVUCSOWORDIQBRZZLD'
         b'BXNAKQAQ9XMVRKMBQTBDFGJZEQBVRLEOUGHZNCJTSZQDDICMLCWZWF9FGAIFDNWUNV'
         b'XMVVXKNCCMRD9NNAKEMILISEKDBYTLPEUXTIGFRWCXKNEOEWBZPHLKPVSNYCEZVNVL'
@@ -737,7 +737,7 @@ class KeyGeneratorTestCase(TestCase):
     self.assertEqual(
       next(generator),
 
-      SigningKey(
+      PrivateKey(
         b'CBKFYOTDYHFWSIOUYAUAHQWOOQDNNQBTSSJPHREUWFBXZFYFHPHZJN9ILAMZYOXLBQ'
         b'MCN9T9F9VSCUSFVQDZROPWBKZ9ENENGFVDLMBNDOAAKZXVBXVAYJLSGVD9ZZSMDCRT'
         b'JUUJPTNRVYYTOALRFOIGAXHUIXWCMXEUGTH9UJKYWTWBEOVWPMNVHJSXXKBAUPYCLS'
