+++ test/crypto/types_test.py
@@ -5,16 +5,131 @@ from __future__ import absolute_import, division, print_function, \
 from unittest import TestCase
 
 from iota import TryteString
-from iota.crypto.types import SigningKey
+from iota.crypto.types import PrivateKey
 
 
 # noinspection SpellCheckingInspection
-class SigningKeyTestCase(TestCase):
+class PrivateKeyTestCase(TestCase):
+  def test_create_signature_fragment(self):
+    """
+    Creating a signature fragment.
+    """
+    key = PrivateKey(
+      b'MPRSBNPSAZRHZSBXSJASXSEHLYVLBXHCUNNQUNYB9CYEVWNBJKZUODPCZGOBRJ9V9C'
+      b'CARPTZEQRXGYDBEZISGMMXGHHHATAUDRTWUHEKJAMKUUKCK9NUUEWEFHPTVAARNLJB'
+      b'XRTFHQDXAVAOWYGJJNLCRVWQFHXPTMLFPGDKEBZCSSOQZEPNRTUKUJ9TRHLWQKOCBZ'
+      b'WBXBQVKUDTKV9IKCDQZRXRGQGIKCQ9BUNYMDFALF9QUFVKHQEHFICORMLLELVXM9VM'
+      b'ZIBUFKHLFGHDCPD9CBZVU9PMXWTAKWCBPBRBUG9PIAYYHKTP9VGDQIP9GUPKK9DVZI'
+      b'UYYTBTBYCDDF9LPBRGNGW9GHHFGRQPYZKKLCUDUAPAKZVIWL9RSJFQO9L9WKMBSWTY'
+      b'LCJJGWTWNLLKMHOUQVKTOUQGEOYNLEHTHJQIPV9ACBEBU9DN9KZWCEZIDEGQYQEYVY'
+      b'WUMPHMJAXOVEXOUMJ9MANMTGBOJAHUGHEYDPUQLNBIRUCLTXRFQIBTEWGUAHIJL9RK'
+      b'LMGKURCYGJLXGFLDF9HPSQUCGSZQDUQGZSYKRKHCCPBAPEBWJUYOBBGUXVSWHRAKWT'
+      b'TMHGJAWM9RPCLMVVNOWNXPBNTPFYWQIPQLNOJJYNRDTLAUBZPLVUACUSPNBHCOPO9Z'
+      b'XSIETGTFDLMRESLN9WXUNQ9SFQFD9CNIETCSVDMNLGNMKOOTOCODTDCPLZPUYON9HD'
+      b'QLDGSCWHKDLQOGDHZHPURB9TSUGVSPKDIZE9ETVMLVMNDRJOOUUFDTSKDDTOGXEVFN'
+      b'BDSBRDZPZTUGYZNVKOIFZYRUIZI9LRIFCCXPHJMHCDOUPKJVUGFPTFI9XPGRDHRZUQ'
+      b'UXCK9MTBINIJFFCQTECFTVTILTKRYJTI9ADGPHFBKNLAPLPKKAOOIEKQLBCDGTZZEF'
+      b'KDBRICXLWZQXXCYVDFAJGKJSS9PADMDCJGJSYSIAOUFHW9CHTDXLVE9PBFCMBZKIHZ'
+      b'9VBCXUDBUUEFNRHI9FXC9KREKFVDGHBFDHNEGCIM9CVSJEQSU9FUKVYSVWCTLM9LAB'
+      b'JCWMPTUHULMGQBVRFCYSGXEQJHMZIVGTIS9QVPTTULFKLSMIPMBUTBXTUQRBOJVUMP'
+      b'BGWEHIFRMNWLHQY9XKFCSGISDWSDILBOSYYWUSICMRWYUWTJY9GW9CC9ZQOKWMIYYD'
+      b'ZRZEKKKJVD9DWTTBEPAJQOVOYPVDGRILVPYXRJZSUXKHCWGVPQNDWKCXJO99AUPVZB'
+      b'VOFELMUKYCRETU9QVUYXXEWHBFFWQOLIHFLRKLGECUMUMHWYESXOUPWLEFMJZVXWJH'
+      b'EBSYVSRSIFXRWBCAEDNUZWEHAKUPAIGIQKCVNURYELBRSKENYWEVPEORWIQHHCSGYW'
+      b'XADSSJGDGTVDUHUNJBPTOTKOS9KKSWSJPRGUB9AFTBBUAMEIBBGIQCCZDRHCCYADMI'
+      b'ATAOUOJOWNUWUKKMYVHNLJSZMTBHYODMJW9RFZNUJCCGTKMVPTBLLUOKXWVJEKSBQW'
+      b'TGNJBEU99LIUBZIHIMX9GHTEXPDMH9OXPFFGMYQFGEJV9TQJPI9ZFVCFZHMRLDVLUG'
+      b'MALNTWOMCNQGHQKZRHSLKFCHDFTOBEOQZRAPDOFEKTQZVRUDCYVZCLONUKMUYIIJRD'
+      b'BVHOZINOYECRMGGIXZNSAGZUOYCNSCEB9EPMZDVMIIBHHLEKSWIDGBOMJRMU9KXKPX'
+      b'DTPULQZNSFKNIFWNPWBRSAQHESXMNCMPIMQLPOKEDFWPYCVYHMXLFFMIQGGBTAM9CP'
+      b'COUJBMZWLYIXOGVFYGLRKSYMPMQBNUMSBV9JXWB9BLJNXOORUTLGI9WSAGFDUVL9DE'
+      b'YGVGYVWUCXXLJQFDGXJWYWNHIUJXFCRWJLVUD9CTLXKZSLGGG9UOKYUX9ESVPNUYLV'
+      b'LGSBVCA99HYBFYZYBEKVSRBFWDURRA9OVXQLKSUAEMKNT9FVYPPRQHCIVPXUS9KP9R'
+      b'OEZ9NOZHYLIURCFY9XFYUEKWQHYKNZEJWKJMFFDJCEWDGVSQLNSQRMJALTGQJIFNED'
+      b'PSYXCIRKYURDVHTNWGRZLDETFXQGWAJTGWAIXXCBNSMMVFCJELLMLEFFYOGNISGWKJ'
+      b'KLAXTGLAYNWEMXUZBKTCFNRRHWMCXMIDPATZCYQFFGNAKTGSIZYPQXDIYUGFCO9YBR'
+      b'LAOEGJUJNL9SHU9NPYQGOLUTDTNQHZPNQGWXXNWYUHLNTCRCOZTVZJQGNSFQLKSNEU'
+      b'LEEYHDMWAWVVHCXVVYNRDLPWDLRGHNPEAOCYESSMABTSMDMPNJHMAKWVEZSQCATTPD'
+      b'IMRA9JOZPVKAZWMEEKTQNQVCRQBVLPCJCKBPNSWFJBM9JBNCUETNUVOHHCZZLMBIOE'
+      b'QEWUOGNIETNOKGFLLKKZUIRDLBAHMZQ9JUWXO9QHZ9XZNAFGQTCJYJUYNKIYTTNQXQ'
+      b'DSUWKEZACQBYRWLTTINDLBRSZPESTWV9KKYCBEEVPKFQIK9RGIADVNDWWMPQOWLYAR'
+      b'VK9BNZLFEZGUHZW9KFZAEDUTVTSGSTGFPCJNX9MGLDFXMTYWXEHOYBTQCVCEXNQNRI'
+      b'YEUGEUJ9AUXMNUTIVYB9VRNVTAAEQTQRPEKYOLECFGPZWWOTIEZMSA9P9QCMCSDMUS'
+      b'HDJOQEUUKFCSHQTWHEHJWQPEXQNKVPNUEMBA9CVOEVJEGJORYWZAOIVWHAMCRVPVTW'
+      b'IMEJNHZ9PBSIPNEUWAQBNSOAWYA9GLPWCNTSAIHLSQD9LJJIVLLVXWSNQHWLYQMQFB'
+      b'WTAUYUGAWOICQJFWTJXN9BLVQDZSTTTGJCCWGWEBGTGODNDCGCSXZMBTSFEZIKMCOZ'
+      b'WJU9IVFAEIGHLCLUNERARLHWQICKGYYEUBBPEEHHHGPTDPEMHMNRYKIXOASAVPN9TT'
+      b'YDEYZHVDGVVTICNCXRMJKJLWJDVPNPQCQESACMEZRDVCFTGMEEHEUFAWZDLNFT9PNI'
+      b'Q9WTWZAMDRYSBSCXIOOYAMQBAHJEFLDUQHPZYDQRPA9MEKAMRBZGTUBRTNYYGBGAY9'
+      b'YIERRWFRMNKGTZSXBONWLGSQVBL9HYSUSUHTVPTIHFFXWDCZSGTBDZQMLPUCYWWVBF'
+      b'9JNJPCQODJYOHXRUTEPRJZW9MJHKQLGIFGQRQWRSBFPASG9YLKOHOKRZQQZPM9MNQC'
+      b'AWAZIU9JRPDXBNMBIYFJUYAEWQSMLSZUPLBKHGIJTNXYQASQUMPMKWIZ9DMINF9CYG'
+      b'ZMILHRJOAJQEOVVALRYQGPPTUOD9OIOHIGRGGJMVKCZUNK9QVBLBTMXJRYWNVY9MZN'
+      b'HZLVJUNZEYICXLLMQSYFPRZQAYJCX9KKFFTPZFCUKRV9EMRHPGPQ9ZNBYXXZXL9LGZ'
+      b'LJZPKVA9HHZEBIIPJYIEBQJO9LIOPTWSUMOLHG9AEJWGANZBFG9IDPFTDDBFYBDPDJ'
+      b'GQEPDFYAYADLQOODRXA9CBUDYXEJWITAKRMBACRDXDCPGGPDCQRHACTJIIVD9XTLVR'
+      b'NNTWCMDTZXXVUOHBOYWCGIFCNEH9AG9KD99LOGNAYIMCRLVTCCCRFWHKVE99AOWOGU'
+      b'UXNITAEYHJHVVOKOVFXVOSXBTGVIVNDJAZITQKXXXLLIMOAKOMZIIXEEYXGNQVYFQQ'
+      b'TJJEDXSDPFA9JZNMWGADIUNVFOURDTGPHCKPEHJMIKZFISIYFOVWWYZYKFRXRH9TCJ'
+      b'GOLCHPYDVFIPRCYLUINFQRBXJ9EDFRJPGPERLTJNTNJMDYHVLYSY9ZALOOEGBNLJQV'
+      b'PAOPVMQPZ9QSUYW9FQFQYKFOYYYPQHIFEWPVWURDY9DWBBVOADLLEFXNLGIFYQXKY9'
+      b'CDDYQXV9EP9MMJW9SUBHL9BGCXYF9LMGUIJDEDJOTSRZZGPKOWHOGFI9QJRNIN9FVB'
+      b'WVFOAJGTYJBOWEPFZDHGKYVGB9IHHCHZODFWRJQSSYDV9IURZBCJALECSLQCXYHAUC'
+      b'GJZNB99QZFZ9FCPFFRXJCUYGWMKRGICUDAUYBKIPPXRQOVBYPKALVCRBQATSBOONFP'
+      b'EMTIYNOSEUC9JLBFMMQTVTBBXVJV9FIOXQDSJCZVTERZHSMTK9JKUYUHRJWOXXTZDD'
+      b'XMJSUTPFQJHKDBMOUWSZRHUJAQOTPXINF9WDGRG99OSMKXGY9WCD9AJVXEDNLWSBWB'
+      b'IHO9ODBQSQL9TUMJZASGNTWUQGNNIWSBNGXRSCLYCGRNTEYTGJAIIXX9YNGHLDZ9QV'
+      b'BCQNMVIQ9EXTSCQAJJROGKIHJJTKLOABYQDHMXJFEKFN9L9FSYAMVOFTXIOSEOBBUA'
+      b'UYSHWINHCJRQTIXWWPZBGKIGJC9TDKFAKZTIFBSDFSXVJCCICIBEAIGNUARDZGGBBH'
+      b'RLMWQ9RSJB9TPLSDEP'
+    )
+
+    self.assertEqual(
+      TryteString.from_trits(
+        key.create_signature(TryteString(b'WGXG9AGGI'), blocks=1)
+      ),
+
+      b'GNCUHGZVIYRQQBXXBUONVL9COKOYDERJAWWI9YRWBVUJLDQQCBMFOORHRTVPKUDFWK'
+      b'YPOOCNVDQOCTJNVCFOENDWEGCXOXZK9BNFGAUTSLGPIWCIWFQAGQTVECGNXLFCUTFQ'
+      b'UJRUQKPJEDUFZTABRPTDXCIE99ZSGZFQCJYRZRSSUEENDBMDRZPLI9KCTVAYTFVRQB'
+      b'HDFLQZJBHIFFDVJPIJFCMOYNKOXQNQHSNDBWXRHLVBXBRLICTY9KJKWK9TZBSNIHKM'
+      b'ADAWFIK9YZVZFZWSZANYEIGQKOVRDGGTGFYFLGKESJKLUJMCKGGAJZHFQJXWIJ9HFI'
+      b'URAXKZKTNULKABTUDTNVQEJRMVUXGOOWAUMSQCUBAOLEFO9PTVPCIFMFJGYAPNFIDW'
+      b'T9TASFHSPXTRVEBRNJOVJIPGMNEWHGZZTQRRHEG9CXBLUJJSCENYKRXKRFPHAOEY9G'
+      b'RNLWUCLXXOBQOSSOVEXZGHBMJKQ9YDQPRGSSNXOFZMUOGKWENWIVIVMDDLPACSKPRS'
+      b'99PNRADWBWSVTXIGGW9THJEQNWEGCWOVSLVWITMUWIUQGSGBYXNKSKUCLFEHWLOOID'
+      b'OKTIOJGPUZQBLHYUKCH9IKPJCRWXBNXGSHIZXYCSMOOQZROSDMNGEMHSKKDOYPKXAI'
+      b'DS9UUINN9RSVSDJUODJ99VEJMLLQRJFPYAZDKTNXPHZYBV9AERYOUJKBXACGZTWWGV'
+      b'NVIDNQPRK9HNEGOVOKQBXOXHZLHFNXDSFJECQVRJFJ9VHSCZACCHRKLWEGE9LAFEQY'
+      b'TFYAGUSYZGFXOIUVATAZCTDMOMVOVMWTDFXRNLHPSYMQKVGRRINYVOBVUXMPZPXBMM'
+      b'OFPNNGDCZHBPQBGMZJKOHOIKTXYZAAFKZEZPTESTUQH9HPNUQKUD9GJSSVDIJQGFRI'
+      b'FULCYSKXTASWYBMXSNJWFFJA9IQOBJYVZVYDXZYMWIYJLXIPGGKULHDBEXYCLQQFXA'
+      b'DJEEVJW9DFWDJQTS9AJWQQZLOYOSQW9DEHHENQVVOKFSBNHSJQHTPGIBLNTCXATNOH'
+      b'TIOQFDIQEZFWWHCUEGNAYAPMUGSNCZQIWGKCYAWZWGVLEZVSZIJKD9EHINYFUVZLPR'
+      b'NIVUNZBVUGMQAUKEDPQWZNKWAJZDVHIFLSQWBGVOXIVBZTYRDBFNXUFFNSSJHJLYJA'
+      b'9MYYLKRFEWMQNYJWYTDNIVNJFCEHBDE99SZTKKGZUQO9AMVGYCIUOCLECKEUBSWBNB'
+      b'9QYSWEUAVROPHIFQPAQUIIPGGTGYSWDFPYYCRFBHBHAUDUX9OKZPIZEOFEIMVHVAKR'
+      b'QJ9HXVC99RNOEOAZVPPIJCWGKCAJOOWRFXSWLAPBNIQGNNKXSBLMNFJCUBYDMPKS9V'
+      b'ASYWPE9WPMZLKSMOFQBPULYKIVBTHTUUNGGRPOOMSSSNUCP9ASWMGONSIANVRPEYLF'
+      b'OPJNSBCXKQSVPECWUWEHIOSVRAKAUTDCPGGMPQXDJFQAAFJQTIDUUFGDMGVZDJCNOS'
+      b'YOEAFJSKXUPHNGDJH9TSKJZGPUETDIFCUQUVARHETQFSRLSBCJCLBR9BJNDTNTWH9P'
+      b'HU99TPAMMYYWJSISCKTBGMROVKWVZTOOIUTOERQKHGPTOZAWKKRGRYLZTJPPUURSMF'
+      b'H9H9EMWIWKJYMWBLNQPFHBUGFKDJHEHORXXDDSFXAAQBSKIORGQNHZYRSUSRNLIBID'
+      b'AVDPZKOLLPOWTCPPUJFTOLMGSPASTNIXBHIJJSBRMREVBCZIPRVBQVPQ9AYYPOVK9L'
+      b'BOQQCGIXAYOXYIC9SYFMGAGZJOIHUZCATPJJOHVAZXLPUVLDJKBXQNPRTWHKH9ZHGC'
+      b'ILKVHYOXNVVWQFMXYDKEAQMXZLNZTFKWRTBTCZLTRCFVQSYG9ZSSZTQJQSHMBOMIQS'
+      b'KNWRG9MEUHWPGAJPJBDKPQOCXDYAIYMUKBOCMTVVZXLPIYQ9XZOJWMICKLECKCRJUX'
+      b'LEYRZCIQNFOIHKJPHTKFTS9YQMBXLCWZBZOWPEWMHLDEHBADWNCSGUYYYNGDGPMJGK'
+      b'MUQ9QCPVPJIOQAZOXMIP9POREGZIDEXJOWXHYTOILXMQBXRGBBNAOIJLRANGZGZSOM'
+      b'XQUEEUWXS9HSNOTUATDNMNIOLERHGBLZN9REGCRXIMGZWTHIGGL9VFGZZFURDCLVCT'
+      b'VKMUTYC9C'
+    )
+
   def test_get_digest_trits(self):
     """
-    Generating digest trits from a valid SigningKey.
+    Generating digest trits from a valid PrivateKey.
     """
-    key = SigningKey(
+    key = PrivateKey(
         b'BWFTZWBZVFOSQYHQFXOPYTZ9SWB9RYYHBOUA9NOYSWGALF9MSVNEDW9A9FLGBRWKED'
         b'MPEIPRKBMRXRLLFJCAGVIMXPISRGXIJQ9BOBHKJEUKDEUUWYXJGCGAWHYBQHBPMRTZ'
         b'FPBGNLMKPZYXZPXFSPFUWZNRWYXUEWMP9URKVVJOSWEPJKSMPLWZPIZGOTVAA9QQOC'
